Refactor wizard related code
- rename sprites package -> wizard
