+++ core/src/main/kotlin/no/ntnu/beardblaster/wizard/Textures.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.sprites
+package no.ntnu.beardblaster.wizard
 
 import com.badlogic.gdx.graphics.g2d.TextureAtlas
