+++ core/src/main/kotlin/no/ntnu/beardblaster/wizard/WizardTemp.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.spell
+package no.ntnu.beardblaster.wizard
 
 class WizardTemp(var remainingHp: Int, var wizardId: String) {
     val maxHp = 30
